+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed a bug in the SchiLD parser that produced an error when trying to import users with more than one space in their username ([#299](https://gitlab.com/Skn0tt/EntE/issues/299))
+
 ## [v0.24.2] - 2019-09-06
 
 Republished `v0.24.0` due to versioning error.